data_project/pyproject.toml
[tool.poetry]
name = "data-project"
version = "0.1.0"
description = ""
authors = ["Egor Dmitriev <egor.dmitriev@luminis.eu>"]

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.118.0"
constructs = "^10.3.0"
sagemaker = "^2.203.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
